DP-91 Add logging aspect; Add string holder for confidential data
Enable automatic logging against a pointcut. This is intended for uniform logging against all application API methods (command and query services) but could also be used for other layers.
Because potential confidential information is passed via the API, a simple string holder is also added, which does not display the string content in its toString() method. This is intended to hold strings passed to API methods like passwords and credit card numbers.

Signed-off-by: Steffan Westcott <steffan.westcott@clicktravel.com>