Set false to exception thrown when new event attributes are added to events
When services that use Cheddar integrate on events the event code is sometimes shared to reduce duplication. When one of these events changes the other must also be built and redeployed because currently a 'UnrecognizedPropertyException' is thrown. The change in settings in this commit to the mapper will stop this behaviour from occurring reducing the coupling between services that integrate on events.