Removed Server side specific error handling when failures occurs while receiving messages
When the SQS queue resource was receiving a message from a queue if an amazon exception occurred it was only retrying if it was of type 'SERVICE'. This turns out to be insufficient as in practice if the host cannot be reached an IOException is thrown which results in a AmazonClientException. This change makes it so that any amazon exceptions result in a logged message and retry just as the requirements stated.