[TRAV-211] changing fetch to use indexes first
Updating the fetch method for DynamoDocumentStoreTemplate so that if the field is indexed the index will be queried rather than a table scan being performed.

Fixed another bug with using comparators on numeric fields.

Added an integration test resulting in some refactoring of DynamoDbTemplateIntegrationTest to take out the DynamoDocumentStoreTemplate test and DynamoDbDataGenerator to be usable by both integration tests.