Support for hash key query
Currently if you try to perform an attribute query on the has value, for example if you want to query hash without range it fails as it attempts to query an index table instead of the table itself.

This change checks that if the query name is the same as the PK because you can't have multiple keys with the same name then it will instead query the table directly. I have noticed an issue with index query around support for multiple keys but i will raise a separate bug report for that.

Signed-off-by: Robin Smith <robin.smith@clicktravel.com>