Fixed silent failure of persistence error when no handler had been defined
If a collection of persistence exception handlers had been passed into the database action class a bug caused no action to be taken if an exception was thrown but did not match any of the handlers. The following code attempts to fix this bug by throwing the original persistence exception if no matching handler has been found and invoked.

Signed-off-by: James Butherway <james.butherway@clicktravel.com>