Remove blue-green deployment code and AWS resource provisioning
The provisioning and deployment models for Cheddar applications have been changed. This commit removes obsolete code supporting the old models and introduces support for the new ones.

Before this commit, Cheddar applications provisioned infrastructure resources such as AWS resources like SQS queues and SNS topics on application startup. Starting with this commit, provisioning of infrastructure is expected to be done before a Cheddar app is deployed. For AWS, CloudFormation is typically used for such tasks. Provisioning is now considered to be out of scope of Cheddar.

Also before this commit, Cheddar had support for blue-green deployments. This attempted to orchestrate REST and message processing from the group of 'old' application instances to the group of 'new' application instances, with a clean switchover such that at no point in time would there be concurrent 'old' and 'new' processing. In practice this proved to be very complex in operation and required extensive support from custom deployment scripts not in Cheddar. Starting with this commit, this approach has been dropped in favour of a standard deployment model which does not prohibit concurrent 'old' and 'new' processing during deployments.

This commit also corrects issues with shutdown of a Cheddar application, as performed when un-deploying an application instance. Cheddar apps now respond to a 'kill' signal by shutting down gracefully, ensuring processing of in-progress REST requests and messages received from queues are completed before the Java process terminates.

Signed-off-by: Steffan Westcott <steffan.westcott@clicktravel.com>