Add test method for published events
This change adds a new test utility method 'InMemoryMessageVerifier#publishedEvents' for use in service tests. This method returns a list of Events that were published using an InMemoryMessagePublisher. The intention is to enable more flexible testing of published events, other than testing for complete equality with existing method 'InMemoryMessageVerifier#eventPublished'.

Signed-off-by: Steffan Westcott <steffan.westcott@clicktravel.com>