Fixed dynamodb fetch bug when dealing with compound GSI with hash the same as tables PK hash
It was discovered that when the hash in a compound GSI was the same collumn as that of the tables hash then instead of querying using the range as well it ignored it and on queried on only the hash value. The fix was fairly benign but the testing was a little more complex but has seen to be working.

Signed-off-by: James Butherway <james.butherway@clicktravel.com>