Don't write null attributes when saving DynamoDB items.
An issue was found where data was being written to a compound GSI without the range part of the key.  This is something which DynamoDB allows but the Cheddar process of converting the item POJOs into JSON was including the null attributes explicitly mapped as null which is not allowed within Dynamo.  This fixes that issue by only writing properties to DynamoDB if they are not null.