Add new message body writer to fix text/plain media type errors
When an endpoint was registering two or more media types in the @Produces annotation, if any were none json such as text/plain, and an error was thrown the mapping was attempting to return a json like object and failing. A new message body writer has been written to allow the correct entity type to be written back in the response. The @Produces annotation has also been removed from the exception mappers which allows the current ‘Accept’ed media type to be passed through. This further enables responses of any json based media types regardless of any specific media type defined on the endpoint.

Signed-off-by: James Butherway <james.butherway@clicktravel.com>