Fixed company id mix up bug in intercom metric controller
When a user was being retreived from intercom its companies where being mapped to the id given by intercom and not the company id which is provided by the creator of the company. This fix should make sure that it is only known company_id’s that mapped back to a User when they are updated.

Signed-off-by: James Butherway <james.butherway@clicktravel.com>