Include a last updated time when the profile is first created
When a user is created in intercom we would like the last seen value to be set as well as the created time.

Without this a user who always uses a privacy tool would never get the last updated value set and as such gets excluded from data syncs etc.

Signed-off-by: Robin <robin.smith@clicktravel.com>