Added AWS XRay support to CHEDDAR
This adds AWS Xray support by manually creating a new segment on each request, storing it in local storage and closing as the response is returned to calling client.

When a segment is open and the AWS clients have been manually instrumented in the service you get output similar to that of what we see in the node projects.

If we are happy with this after letting it soak UAT with Wire (PR coming next) we can look into how to have this only turned on in production.